# services/risk-service/src/main/java/br/com/orquestrapay/risk/api/RespostaAnaliseRisco.java
package br.com.orquestrapay.risk.api;

import java.time.Instant;
import java.util.UUID;

public record RespostaAnaliseRisco(
        UUID idAnalise,
        UUID idCompra,
        int pontuacao,
        boolean aprovada,
        String sinais,
        String modeloDecisao,
        String versaoModeloDecisao,
        Instant analisadaEm) {
}